feat(web): 共通コンポーネント抽出 (StatusBadge / TerminalOutput) #11
- StatusBadge: pill (inline-style) と tag (Tailwind) の2バリアント対応
- TerminalOutput: auto-scroll内蔵 + 外部ref注入に対応
- design-tokens.ts: STATUS_COLORS / STATUS_STYLES を一元管理
- runs, runs/[runId], live, monitor の4ページからインライン定義を削除しimportに置換
